Extract topic store lifecycle
